+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Added support for chat message popout for item chat cards
 - Added more localization to Item Chat Cards
+- Added ability to assign any card to a combatant via combatant config
 
 ### Changed